pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ctf"
version = "0.1.0"
description = "Local LangGraph runtime for ctf"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docstring-parser>=0.17.0",
    "fastmcp>=2.0.0",
    "gql>=4.0.0",
    "ipykernel>=6.29.0",
    "jupyter-client>=8.6.3",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.3.0",
    "libtmux>=0.47.0",
    "mcp>=1.0.0",
    "nbformat>=5.10.4",
    "playwright>=1.55.0",
    "psutil>=7.1.3",
    "pydantic>=2.12.4",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "requests-toolbelt>=1.0.0",
]

[tool.setuptools]
py-modules = ["main"]
packages = ["runtime"]
